+++ src/test/java/test/thread/SequentialTest.java
@@ -77,9 +77,9 @@ public class SequentialTest extends BaseTest {
     }
 
     long[] ids = new long[] {
-        maps[0].keySet().iterator().next().longValue(),
-        maps[1].keySet().iterator().next().longValue(),
-        maps[2].keySet().iterator().next().longValue(),
+            maps[0].keySet().iterator().next(),
+            maps[1].keySet().iterator().next(),
+            maps[2].keySet().iterator().next(),
     };
     Map<Long, Long> verifyMap = new HashMap<>();
